CASMPET-4135 Increase TTLs
This increases the TTL of the ckdump workloads to 10 days. The CA TTL is
also increased to 90 days to be able to handle longer TTLs. The CA TTL
needs to be 6 times the largest TTL. This allows us to increase the TTL
of the kdump tokens to 15 days if needed.