CASMINST-3919: Add post-upgrade hook to set pooler pod antiaffinity
Upgrades were hitting problems when all of the pooler pods wound
up on the same node. While the postgres operator is enhanced to
set the anti-affinity on the pooler deployment it's not smart
enough to update the deployment if it already exists. This adds
a post-upgrade hook that just updates the deployment with the
same settings that the postgres operator would set if it created
the deployment.