CASMINST-3917: Set pod anti-affinity for spire-jwks
The spire-jwks pods could all be running on the same node, and if
that goes down then spire-jwks is not available and this causes
requests through the API gateway to fail due to the auth check in
OPA to fail.

The proposed fix is to set pod-antiaffinity for spire-jwks so that
the pods will not all wind up on the same node.